Add CRUD for dynamic

+++ src/main/java/com/zhytnik/shop/domain/dynamic/PrimitiveType.java
@@ -5,5 +5,5 @@ package com.zhytnik.shop.domain.dynamic;
  * @since 28.05.2016
  */
 public enum PrimitiveType {
-    BOOLEAN, DOUBLE, INTEGER, DATE, STRING, GALLERY, MAP, CURRENCY
+    BOOLEAN, DOUBLE, LONG, DATE, STRING, GALLERY, MAP, CURRENCY
 }
